+++ starkbank/__init__.py
@@ -1,4 +1,4 @@
-version = "2.28.1"
+version = "2.29.0"
 
 user = None
 language = "en-US"